Move octal to decimal function to Converter.
- Octal to decimal function tests moved to ConverterTest.
